Extract new color helper class
